+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.11] - 2025-05-31
+
+### Removed
+
+- Removed SHA2 functions (sha256Async, sha384Async, sha512Async, sha2Async)
+
+### Changed
+
+- Updated expo-crypto-universal dependency to version 0.2.11
+
 ## [0.2.9] - 2025-05-31
 
 ### Changed